+++ library/src/device/solution-shipping.py
@@ -296,8 +296,8 @@ def cli():
     archs = args.gpu_arch.split(' ')
     if 'all' in archs:
         archs = [
-            'gfx900', 'gfx906', 'gfx908', 'gfx90a', 'gfx1030', 'gfx1100',
-            'gfx1101', 'gfx1102'
+            'gfx900', 'gfx906', 'gfx908', 'gfx90a', 'gfx940', 'gfx941',
+            'gfx942', 'gfx1030', 'gfx1100', 'gfx1101', 'gfx1102'
         ]
 
     # remove xnack and sramecc