+++ library/src/include/twiddles.h
@@ -37,13 +37,15 @@ gpubuf twiddles_create(size_t                     N,
                        bool                       attach_halfN,
                        const std::vector<size_t>& radices,
                        unsigned int               deviceId);
-gpubuf twiddles_create_2D(size_t                 N1,
-                          size_t                 N2,
-                          rocfft_precision       precision,
-                          const hipDeviceProp_t& deviceProp,
-                          bool                   attach_halfN,
-                          bool                   attach_halfN2,
-                          unsigned int           deviceId);
+gpubuf twiddles_create_2D(size_t                     N1,
+                          size_t                     N2,
+                          rocfft_precision           precision,
+                          const hipDeviceProp_t&     deviceProp,
+                          bool                       attach_halfN,
+                          bool                       attach_halfN2,
+                          const std::vector<size_t>& radices1,
+                          const std::vector<size_t>& radices2,
+                          unsigned int               deviceId);
 
 void twiddle_streams_cleanup();