+++ library/src/repo.cpp
@@ -192,21 +192,30 @@ std::pair<void*, size_t> Repo::GetTwiddles1D(size_t                     length,
         });
 }
 
-std::pair<void*, size_t> Repo::GetTwiddles2D(size_t                 length0,
-                                             size_t                 length1,
-                                             rocfft_precision       precision,
-                                             const hipDeviceProp_t& deviceProp,
-                                             bool                   attach_halfN1,
-                                             bool                   attach_halfN2)
+std::pair<void*, size_t> Repo::GetTwiddles2D(size_t                     length0,
+                                             size_t                     length1,
+                                             rocfft_precision           precision,
+                                             const hipDeviceProp_t&     deviceProp,
+                                             bool                       attach_halfN1,
+                                             bool                       attach_halfN2,
+                                             const std::vector<size_t>& radices1,
+                                             const std::vector<size_t>& radices2)
 {
     std::lock_guard<std::mutex> lck(mtx);
     Repo&                       repo = Repo::GetRepo();
 
-    repo_twd_key_2D_t key{length0, length1, precision};
+    repo_twd_key_2D_t key{length0, length1, precision, radices1, radices2};
     return GetTwiddlesInternal(
         key, repo.twiddles_2D, repo.twiddles_2D_reverse, [&](unsigned int deviceId) {
-            return twiddles_create_2D(
-                length0, length1, precision, deviceProp, attach_halfN1, attach_halfN2, deviceId);
+            return twiddles_create_2D(length0,
+                                      length1,
+                                      precision,
+                                      deviceProp,
+                                      attach_halfN1,
+                                      attach_halfN2,
+                                      radices1,
+                                      radices2,
+                                      deviceId);
         });
 }