+++ library/src/tree_node.cpp
@@ -119,19 +119,11 @@ bool LeafNode::KernelCheck(std::vector<FMKey>& kernel_keys)
         FMKey assignedKey = kernel_keys.front();
         kernel_keys.erase(kernel_keys.begin());
 
-        // for kernels that haven't been supported tuning. (i.e. 2D_SINGLE)
-        // don't assign specified_key
-        if(assignedKey == FMKey::EmptyFMKey())
-        {
-            if(LOG_TRACE_ENABLED())
-                (*LogSingleton::GetInstance().GetTraceOS())
-                    << "kernel is not tuned, use default kernel." << std::endl;
-        }
         // check if the assigned key is consistent with the node information
-        else if((length[0] != assignedKey.lengths[0])
-                || (dimension == 2 && length[1] != assignedKey.lengths[1])
-                || (precision != assignedKey.precision) || (scheme != assignedKey.scheme)
-                || (ebtype != assignedKey.kernel_config.ebType))
+        if((length[0] != assignedKey.lengths[0])
+           || (dimension == 2 && length[1] != assignedKey.lengths[1])
+           || (precision != assignedKey.precision) || (scheme != assignedKey.scheme)
+           || (ebtype != assignedKey.kernel_config.ebType))
         {
             if(LOG_TRACE_ENABLED())
                 (*LogSingleton::GetInstance().GetTraceOS())