+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Full documentation for rocFFT is available at [rocm.docs.amd.com](https://rocm.d
 ## rocFFT 1.0.25 for ROCm 6.0.0
 
 ### Optimizations
-- Improved performance of some small 2D/3D real FFTs supported by 2D_SINGLE kernel.
+- Improved performance of some small 2D/3D real FFTs supported by 2D_SINGLE kernel. gfx90a gets more optimization
+  by offline tuning.
 - Removed an extra kernel launch from even-length real-complex FFTs that use callbacks.
 
 ### Changed